+++ docs/development/releasenotes.rst
@@ -1,6 +1,14 @@
 Release Notes
 =============
 
+v0.20
+-----
+
+Tags
+~~~~
+1. *Tables*. Tags to render tables.
+2. *Data*. Tags to incorporate data into tables or plots.
+
 v0.19
 -----
 
@@ -12,6 +20,7 @@ Tags
    tag by default. Instead, a label tag is created, and the content of the tag
    is included if a ``@label.title`` macro was included.
 3. *Equations*
+
    3.1. *tex*: Block equations (equations in their own paragraph) are written
         without newlines before/after the environment in latex.